Move Boundary classes formulas as util
