+++ iidm/iidm-api/src/main/java/com/powsybl/iidm/network/DanglingLine.java
@@ -6,6 +6,8 @@
  */
 package com.powsybl.iidm.network;
 
+import com.powsybl.iidm.network.util.DanglingLineBoundaryImpl;
+
 /**
  * A dangling line to model boundaries (X nodes).
  * <p>A dangling line is a component that aggregates a line chunk and a constant
@@ -279,6 +281,6 @@ public interface DanglingLine extends Injection<DanglingLine>, FlowsLimitsHolder
     CurrentLimitsAdder newCurrentLimits();
 
     default Boundary getBoundary() {
-        throw new UnsupportedOperationException();
+        return new DanglingLineBoundaryImpl(this);
     }
 }
