+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/HalfLineBoundaryImpl.java
@@ -6,76 +6,21 @@
  */
 package com.powsybl.iidm.network.impl;
 
-import com.powsybl.iidm.network.*;
-import com.powsybl.iidm.network.Boundary;
-import com.powsybl.iidm.network.util.SV;
-
-import java.util.Objects;
+import com.powsybl.iidm.network.Branch;
+import com.powsybl.iidm.network.TieLine;
+import com.powsybl.iidm.network.util.AbstractHalfLineBoundaryImpl;
 
 /**
  * @author Miora Ralambotiana <miora.ralambotiana at rte-france.com>
  */
-class HalfLineBoundaryImpl implements Boundary {
-    // side represents the network side.
-    // side here is Side.ONE for the half line 1 of a tie line.
-    // side is Side.TWO for the half line 2 of a tie line.
-
-    private final TieLineImpl.HalfLineImpl parent;
-    private final Branch.Side side;
+class HalfLineBoundaryImpl extends AbstractHalfLineBoundaryImpl {
 
     HalfLineBoundaryImpl(TieLineImpl.HalfLineImpl parent, Branch.Side side) {
-        this.parent = Objects.requireNonNull(parent);
-        this.side = Objects.requireNonNull(side);
-    }
-
-    @Override
-    public double getV() {
-        Terminal t = getConnectable().getTerminal(side);
-        Bus b = t.getBusView().getBus();
-        return new SV(t.getP(), t.getQ(), getV(b), getAngle(b), side).otherSideU(parent);
-    }
-
-    @Override
-    public double getAngle() {
-        Terminal t = getConnectable().getTerminal(side);
-        Bus b = t.getBusView().getBus();
-        return new SV(t.getP(), t.getQ(), getV(b), getAngle(b), side).otherSideA(parent);
-    }
-
-    @Override
-    public double getP() {
-        Terminal t = getConnectable().getTerminal(side);
-        Bus b = t.getBusView().getBus();
-        return new SV(t.getP(), t.getQ(), getV(b), getAngle(b), side).otherSideP(parent);
-    }
-
-    @Override
-    public double getQ() {
-        Terminal t = getConnectable().getTerminal(side);
-        Bus b = t.getBusView().getBus();
-        return new SV(t.getP(), t.getQ(), getV(b), getAngle(b), side).otherSideQ(parent);
-    }
-
-    @Override
-    public Branch.Side getSide() {
-        return side;
+        super(parent, side);
     }
 
     @Override
     public TieLine getConnectable() {
-        return parent.getParent();
-    }
-
-    @Override
-    public VoltageLevel getVoltageLevel() {
-        return getConnectable().getTerminal(side).getVoltageLevel();
-    }
-
-    private static double getV(Bus b) {
-        return b == null ? Double.NaN : b.getV();
-    }
-
-    private static double getAngle(Bus b) {
-        return b == null ? Double.NaN : b.getAngle();
+        return ((TieLineImpl.HalfLineImpl) getParent()).getParent();
     }
 }
